:bug: Fix protocol & timing issues w/ bit_bang_i2c
The state of the output pins provided to `hal::bit_bang_i2c` can be in
any state, meaning that the i2c bus may be in a state that will not work
initially. In order to initialize the bus to a known state,
`hal::probe(*this, 0x00)` is called twice.